requirements.txt
# scraping
beautifulsoup4
requests
docling

# fastapi
fastapi
uvicorn
python-multipart

# AI / RAG
openai
chromadb
sentence-transformers

# firebase
firebase-admin

# environment variables
python-dotenv

# logging
loguru
